backend/requirements.txt
aioredis==1.3.1
aiosignal==1.3.1
altair==4.2.2
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==24.4.2
Automat==25.4.16
autopep8==2.3.2
bcrypt==5.0.0
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.3
celery==5.3.6
certifi==2023.11.17
cffi==2.0.0
channels==3.0.5
channels-redis==3.4.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
CMRESHandler==1.0.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.1
curlify==2.2.1
daphne==3.0.2
defusedxml==0.7.1
Django==3.2
django-dbbackup==3.3.0
django-filter==23.5
django-redis==5.4.0
django-silk==5.2.0
djangorestframework==3.14.0
djangorestframework-bulk==0.2.1
djangorestframework-simplejwt==5.3.1
docx2pdf==0.1.8
drf-spectacular==0.28.0
elastic-transport==8.11.0
elasticsearch==7.12.0
elasticsearch-dsl==7.4.0
entrypoints==0.4
et-xmlfile==1.1.0
faiss-cpu==1.13.0
Faker==37.8.0
filelock==3.20.0
frozenlist==1.4.1
fsspec==2025.10.0
gevent==21.12.0
gitdb==4.0.11
GitPython==3.1.42
gprof2dot==2025.4.14
greenlet==1.1.3.post0
hiredis==3.2.1
huggingface-hub==0.36.0
hyperlink==21.0.0
idna==3.6
importlib_metadata==7.0.2
incremental==24.7.2
inflection==0.5.1
invoke==2.2.1
Jinja2==3.1.3
jira==3.6.0
joblib==1.5.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.4
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.0.4
networkx==3.4.2
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==1.5.3
paramiko==4.0.0
pillow==10.2.0
plum-dispatch==1.7.4
prompt-toolkit==3.0.43
protobuf==3.20.3
pyarrow==15.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pycryptodome==3.20.0
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
Pympler==1.0.1
PyMySQL==1.1.0
PyNaCl==1.6.1
pyOpenSSL==25.3.0
python-dateutil==2.8.2
python-docx==1.1.2
python-json-logger==3.3.0
pytz==2023.3.post1
pywin32==311
PyYAML==6.0.1
qrcode==8.2
redis==3.5.3
redis-py-cluster==2.1.3
referencing==0.33.0
regex==2025.11.3
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.1
rocketmq==0.4.4
rocketmq-client-python==2.0.0
rpds-py==0.18.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
service-identity==24.2.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sqlparse==0.4.4
sympy==1.14.0
tenacity==8.2.3
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.2.1
toolz==0.12.1
torch==2.9.1
tornado==6.4
tqdm==4.66.4
transformers==4.57.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.2.0
urllib3==1.26.18
validators==0.22.0
vine==5.1.0
watchdog==4.0.0
wcwidth==0.2.13
yarl==1.9.4
zipp==3.17.0
zope.event==6.0
zope.interface==8.0
